+++ src/router/routes/specialRoutes.ts
@@ -5,14 +5,9 @@ import { views } from "./routesMaps";
 
 /* ========== */
 
-export const specialRoutes: Array<RouteRecordRaw> = [
+// 重定向到原创角色
+const oc: Array<RouteRecordRaw> = [
     {
-        // 重定向到首页
-        path: "/",
-        redirect: "/index",
-    },
-    {
-        // 重定向到原创角色
         path: "/oc",
         redirect: "/original-character",
     },
@@ -21,16 +16,67 @@ export const specialRoutes: Array<RouteRecordRaw> = [
         path: "/oc/:path(.*)",
         redirect: to => `/original-character/${to.params.path}`,
     },
+];
+
+// 重定向到感谢页面
+const thanks: Array<RouteRecordRaw> = [
     {
-        // 重定向到致谢页面
+        path: "/thank",
+        redirect: "/acknowledgments",
+    },
+    {
+        path: "/thank/:path(.*)",
+        redirect: to => `/acknowledgments/${to.params.path}`,
+    },
+    {
         path: "/thanks",
         redirect: "/acknowledgments",
     },
     {
-        // 致谢子页面路由转发
         path: "/thanks/:path(.*)",
         redirect: to => `/acknowledgments/${to.params.path}`,
     },
+];
+
+// 重定向到文档页面
+const document: Array<RouteRecordRaw> = [
+    {
+        path: "/doc",
+        redirect: "/documents",
+    },
+    {
+        path: "/doc/:path(.*)",
+        redirect: to => `/documents/${to.params.path}`,
+    },
+    {
+        path: "/docs",
+        redirect: "/documents",
+    },
+    {
+        path: "/docs/:path(.*)",
+        redirect: to => `/documents/${to.params.path}`,
+    },
+    {
+        path: "/document",
+        redirect: "/documents",
+    },
+    {
+        path: "/document/:path(.*)",
+        redirect: to => `/documents/${to.params.path}`,
+    },
+];
+
+/* ========== */
+
+export const specialRoutes: Array<RouteRecordRaw> = [
+    {
+        // 重定向到首页
+        path: "/",
+        redirect: "/index",
+    },
+    ...oc,
+    ...thanks,
+    ...document,
     {
         // 404 页面
         path: "/:pathMatch(.*)*",