+++ src/views/DocumentPage.vue
@@ -1,0 +1,9 @@
+<template>
+    <section class="document-page">
+        <RouterView />
+    </section>
+</template>
+
+<script setup lang="ts">
+import { RouterView } from "vue-router";
+</script>