+++ src/views/ThanksPage.vue
@@ -15,6 +15,7 @@
 <script setup lang="ts">
 import MobileAccessReminder from "@/components/ui/MobileAccessReminder.vue";
 import thanksDocs from "@/assets/docs/Acknowledgments.md";
+import * as thanksData from "@/assets/data/thanks";
 
 const lastUpdated: string = "2025-10-04";
 const version: string = "1.0";
@@ -26,10 +27,10 @@ const version: string = "1.0";
 .page-footer {
     .flex-end-center();
     color: @text-muted;
-    font-size: 0.9rem;
+    font-size: 14.4px;
 
     .separator {
-        margin: 0 0.5rem;
+        margin: 0 8px;
     }
 }
 </style>